% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalue.R
\name{calculate_pvalue}
\alias{calculate_pvalue}
\title{Corrected Wilcoxon tests}
\usage{
calculate_pvalue(data, p_threshold = 0.05)
}
\arguments{
\item{data}{dataframe containing columns named high (logical),
cell_type (factor) and value (float).}

\item{p_threshold}{float for the significativity threshold of the P-value.}
}
\value{
rstatix_test object for a table with cell types in the row and
P-values, corrections and other statistics in the column.
}
\description{
Displays stars for each cell type corresponding to the significance level
of Wilcoxon tests between expression levels (high or low) with
Benjamini-Hochberg multiple correction.
}
\examples{
data(tcga)
df <- convert_biodata(tcga$genes, tcga$cells, "A")
stats <- calculate_pvalue(df)
}
