% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_biodata.R
\name{convert_biodata}
\alias{convert_biodata}
\title{Format biological data}
\usage{
convert_biodata(gene, cell_type, select = colnames(gene)[3], stat = "mean")
}
\arguments{
\item{gene}{data frame whose first two columns contain identifiers and the
others float values.}

\item{cell_type}{data frame whose first two columns contain identifiers and
the others float values.}

\item{select}{character for a column name in gene.}

\item{stat}{character for the statistic to be chosen among "mean", "median"
or "quantile".}
}
\value{
data frame with the following columns:
\itemize{
\item \code{high} (logical): the expression levels of a selected gene,
TRUE for below or FALSE for above average.
\item \code{cell_type} (factor): cell types.
\item \code{value} (float): the abundance estimation of the cell types.
}
}
\description{
Merges gene and cell datasets with the same TCGA sample identifiers,
splits samples according to the expression levels of a selected gene into
two categories (below or above average) and formats into a 3-column
data frame: gene expression levels, cell types, and gene expression values.
}
\examples{
data(tcga)
(df_formatted <- convert_biodata(tcga$genes, tcga$cells, "A"))
}
