% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plot expression}
\usage{
plot_violin(
  data,
  gene = "Gene X",
  type = "violin",
  title = NULL,
  label = "value",
  ...
)
}
\arguments{
\item{data}{dataframe containing columns named high (logical),
cell_type (factor) and value (float).}

\item{gene}{character for the name of the gene plotted in the title.}

\item{type}{character for the type of plot to be chosen among "violin"
or "boxplot".}

\item{title}{character for the title of the plot.}

\item{label}{character for the title of the x-axis.}

\item{...}{arguments to pass to ggplot::theme().}
}
\value{
ggrub object for a violon plot.
}
\description{
Violin plot of cell subtypes according to the expression level (high or
low) of a selected gene.
}
\examples{
library(ggpubr)
data(tcga)
df <- convert_biodata(tcga$genes, tcga$cells, "A")
p <- plot_violin(df)
stats <- calculate_pvalue(df)
p + stat_pvalue_manual(stats, label = "p.adj.signif")
}
