% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\docType{data}
\name{tcga}
\alias{tcga}
\title{Random biological data}
\format{
A list of 3 data frames each containing 10 individuals and 5
numeric variables.
}
\source{
\url{https://cancergenome.nih.gov/} \cr
\url{https://cibersort.stanford.edu/}
}
\usage{
data(tcga)
}
\description{
A list of randomly generated biological data: RNASeq data, phenotypic
metadata and cell abundance.
}
\details{
\itemize{
\item \code{genes}: Random data based on the RNASeq distribution of The
Cancer Genome Atlas (TCGA) database.
\item \code{phenotypes}: Metadata based on the breast cancer model from
the TCGA database containing sample ID, sample type ID, sample type and
primary disease.
\item \code{cells}: Random data based on the distribution of abundance
estimates of certain cell types using the Cibersort_ABS algorithm.
}
}
\examples{
library(ggpubr)
data(tcga)
df <- (convert2biodata(
algorithm = "Cibersort_ABS",
disease = "breast invasive carcinoma",
tissue = "Primary Tumor",
gene_x = "A"
))
stats <- calculate_pvalue(df)
plot_violin(df, stats = stats)
}
\keyword{datasets}
