% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaViz-package.R
\docType{package}
\name{tcgaViz-package}
\alias{tcgaViz}
\alias{tcgaViz-package}
\title{tcgaViz: Vizualisation tool for The Cancer Genome Atlas Program (TCGA)}
\description{
Differential analysis of tumor tissue immune cell type
    abundance based on RNASeq gene-level expression from The Cancer Genome
    Atlas (TCGA) database.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Gilles Marodon
  \item Nicolas Aubert
}

}
\keyword{internal}
