% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaViz-package.R
\docType{package}
\name{tcgaViz-package}
\alias{tcgaViz}
\alias{tcgaViz-package}
\title{tcgaViz: Vizualisation tool for The Cancer Genome Atlas Program (TCGA)}
\description{
A vizualisation tool to depict proportions of immune cells in any tumor relative to the average expression of any gene (or signature).
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Gilles Marodon
  \item Nicolas Petit
}

}
\keyword{internal}
