% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaViz-package.R
\docType{package}
\name{tcgaViz-package}
\alias{tcgaViz}
\alias{tcgaViz-package}
\title{tcgaViz: Visualization Tool for the Cancer Genome Atlas Program (TCGA)}
\description{
Differential analysis of tumor tissue immune cell type
    abundance based on RNA-seq gene-level expression from The Cancer
    Genome Atlas (TCGA; <https://pancanatlas.xenahubs.net>) database.
}
\seealso{
\itemize{
\item \href{https://github.com/ecamenen/tcgaViz}{Contribute to the development}
\item \href{https://github.com/ecamenen/tcgaViz/issues}{Report bugs}
\item \href{https://github.com/ecamenen/tcgaViz/blob/master/doc/Tutorial.pdf}{Read the tutorial}
}
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Gilles Marodon
  \item Nicolas Aubert
}

}
\keyword{internal}
